Message table in database changes the primary key to non auto-Increment. (#180)

+++ src/DotNetCore.CAP/LoggerExtensions.cs
@@ -10,12 +10,12 @@ namespace DotNetCore.CAP
     [SuppressMessage("ReSharper", "InconsistentNaming")]
     internal static class LoggerExtensions
     {
-        public static void ConsumerExecutedAfterThreshold(this ILogger logger, int messageId, int retries)
+        public static void ConsumerExecutedAfterThreshold(this ILogger logger, long messageId, int retries)
         {
             logger.LogWarning($"The Subscriber of the message({messageId}) still fails after {retries}th executions and we will stop retrying.");
         }
 
-        public static void SenderAfterThreshold(this ILogger logger, int messageId, int retries)
+        public static void SenderAfterThreshold(this ILogger logger, long messageId, int retries)
         {
             logger.LogWarning($"The Publisher of the message({messageId}) still fails after {retries}th sends and we will stop retrying.");
         }
@@ -25,12 +25,12 @@ namespace DotNetCore.CAP
             logger.LogWarning(ex, "FailedThresholdCallback action raised an exception:" + ex.Message);
         }
 
-        public static void ConsumerExecutionRetrying(this ILogger logger, int messageId, int retries)
+        public static void ConsumerExecutionRetrying(this ILogger logger, long messageId, int retries)
         {
             logger.LogWarning($"The {retries}th retrying consume a message failed. message id: {messageId}");
         }
 
-        public static void SenderRetrying(this ILogger logger, int messageId, int retries)
+        public static void SenderRetrying(this ILogger logger, long messageId, int retries)
         {
             logger.LogWarning($"The {retries}th retrying send a message failed. message id: {messageId} ");
         }
@@ -40,7 +40,7 @@ namespace DotNetCore.CAP
             logger.LogDebug($"Message published. name: {name}, content:{content}.");
         }
 
-        public static void MessagePublishException(this ILogger logger, int messageId, string reason, Exception ex)
+        public static void MessagePublishException(this ILogger logger, long messageId, string reason, Exception ex)
         {
             logger.LogError(ex, $"An exception occured while publishing a message, reason:{reason}. message id:{messageId}");
         }
